Refactoring of old plugins (Model_Original + Extract_Align) + Cleanups
