+++ faceswap.py
@@ -5,7 +5,7 @@ if  sys.version_info[0] < 3:
 if sys.version_info[0] == 3 and sys.version_info[1] < 2:
     raise Exception("This program requires at least python3.2")
 
-from lib.utils import FullHelpArgumentParser
+from lib.cli import FullHelpArgumentParser
 
 from scripts.extract import ExtractTrainingData
 from scripts.train import TrainingProcessor
