+++ plugins/Extract_Align/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+__author__ = """Based on https://reddit.com/u/deepfakes/"""
+__version__ = '0.1.0'
+
+from .Extract import Extract
