+++ plugins/Model_Original/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+__author__ = """Based on https://reddit.com/u/deepfakes/"""
+__version__ = '0.1.0'
+
+from .Model import Model
+from .Trainer import Trainer
+from .AutoEncoder import AutoEncoder
